backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
deepface==0.0.79
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
tf-keras==2.16.0
tensorflow==2.16.1